Rename X to XBrowser for clarity based on feedback.
