+++ src/frontend/src/routes/(new-styling)/unsupported/+page.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-  import X from "./components/views/X.svelte";
+  import XBrowser from "./components/views/XBrowser.svelte";
   import Generic from "./components/views/Generic.svelte";
 
   const isX = /\bTwitter/i.test(navigator.userAgent);
 </script>
 
 {#if isX}
-  <X />
+  <XBrowser />
 {:else}
   <Generic />
 {/if}
