+++ src/frontend/src/routes/(new-styling)/unsupported/components/views/XBrowser.svelte
